finished bonus

+++ routes/notes.js
@@ -6,12 +6,40 @@ const router = express.Router();
 const mongoose = require('mongoose');
 
 const Note = require('../models/note');
+const Folder = require('../models/folder');
+const Tag = require('../models/tag');
+
+//validate Folder
+function validateFolder(userId, folderId) {
+  if (!folderId) {
+    return Promise.resolve();
+  }
+  return Folder.findOne({_id: folderId, userId})
+    .then(result => {
+      if (!result) {
+        return Promise.reject('InvalidFolder');
+      }
+    });
+}
+
+//validate Tags
+function validateTagIds(userId, tags = []) {
+  if (!tags.length) {
+    return Promise.resolve();
+  }
+  return Tag.find({ $and: [{ _id: { $in: tags }, userId }] })
+    .then(results => {
+      if (tags.length !== results.length) {
+        return Promise.reject('InvalidTag');
+      }
+    });
+}
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/notes', (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
-
-  let filter = {};
+  const userId = req.user.id;
+  let filter = { userId};
 
   /**
    * BONUS CHALLENGE - Search both title and content using $OR Operator
@@ -45,6 +73,7 @@ router.get('/notes', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -52,7 +81,7 @@ router.get('/notes/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findById({ _id: id, userId })
     .populate('tags')
     .then(result => {
       if (result) {
@@ -69,7 +98,8 @@ router.get('/notes/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/notes', (req, res, next) => {
   const { title, content, folderId, tags } = req.body;
-
+  const userId = req.user.id;
+  const newItem = { title, content, tags, userId };
   /***** Never trust users - validate input *****/
   if (!title) {
     const err = new Error('Missing `title` in request body');
@@ -77,23 +107,23 @@ router.post('/notes', (req, res, next) => {
     return next(err);
   }
 
-  if (tags) {
-    tags.forEach((tag) => {
-      if (!mongoose.Types.ObjectId.isValid(tag)) {
-        const err = new Error('The `id` is not valid');
-        err.status = 400;
-        return next(err);
-      }
-    });
+  if(mongoose.Types.ObjectId.isValid(folderId)) {
+    newItem.folderId = folderId;
   }
 
-  const newItem = { title, content, folderId, tags };
+  const valFolderIdProm = validateFolder(userId, folderId);
+  const valTagIdsProm = validateTagIds(userId, tags);
 
-  Note.create(newItem)
+  Promise.all([valFolderIdProm, valTagIdsProm])
+    .then(() => Note.create(newItem))
     .then(result => {
       res.location(`${req.originalUrl}/${result.id}`).status(201).json(result);
     })
     .catch(err => {
+      if(err === 'InvalidFolder') {
+        err = new Error('Folder not valid');
+        err.status = 400;
+      }
       next(err);
     });
 });
@@ -102,6 +132,9 @@ router.post('/notes', (req, res, next) => {
 router.put('/notes/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags } = req.body;
+  const userId = req.user.id;
+  const updateNote = { title, content, tags, userId };
+
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -117,25 +150,17 @@ router.put('/notes/:id', (req, res, next) => {
   }
 
   if (mongoose.Types.ObjectId.isValid(folderId)) {
-    updateItem.folderId = folderId;
+    updateNote.folderId = folderId;
   }
 
-  if (tags) {
-    tags.forEach((tag) => {
-      if (!mongoose.Types.ObjectId.isValid(tag)) {
-        const err = new Error('The `id` is not valid');
-        err.status = 400;
-        return next(err);
-      }
-    });
-  }
+  const valFolderIdProm = validateFolder(userId, folderId);
+  const valTagIdsProm = validateTagIds(userId, tags);
 
-
-  const updateItem = { title, content, tags };
-  const options = { new: true };
-
-  Note.findByIdAndUpdate(id, updateItem, options)
-    .populate('tags')
+  Promise.all([valFolderIdProm, valTagIdsProm])
+    .then(() => {
+      return Note.findByIdAndUpdate(id, updateNote, {new: true})
+        .populate('tags');
+    })
     .then(result => {
       if (result) {
         res.json(result);
@@ -144,6 +169,14 @@ router.put('/notes/:id', (req, res, next) => {
       }
     })
     .catch(err => {
+      if (err === 'InvalidFolder') {
+        err = new Error('The folder is not valid');
+        err.status = 400;
+      }
+      if (err === 'InvalidTag') {
+        err = new Error('The tag is not valid');
+        err.status = 400;
+      }
       next(err);
     });
 });
@@ -151,10 +184,15 @@ router.put('/notes/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
-  Note.findByIdAndRemove(id)
-    .then(() => {
-      res.status(204).end();
+  Note.findByIdAndRemove({ _id: id, userId })
+    .then(result => {
+      if (result) {
+        res.status(204).end();
+      } else {
+        next();
+      }
     })
     .catch(err => {
       next(err);
